fix(core): keep daemon.ts free of the backup engine's runtime import graph
NodeDashboard imports daemon.ts (readPidFile), and daemon.ts's value imports
from backup-host.js were dragging ApiClient/ApiError/better-sqlite3 into
every consumer's module graph — breaking the two TUI dashboard suites that
mock '../api.js' partially. BACKUP_HOST_EVENT + BackupHostEventFrame now
live in backup/events.ts (dependency-free, re-exported from backup-host for
existing consumers), daemon.ts imports BackupHost as type-only, and the busy
rejection is duck-typed on err.code instead of instanceof.

Relates to #318 (epic #308)
